+++ Sources/MapboxNavigation/SpriteRepository.swift
@@ -4,7 +4,11 @@ import MapboxCoreNavigation
 import MapboxDirections
 
 class SpriteRepository {
-    let imageCache = ImageCache()
+    // Caching the Sprite and single shield icon images.
+    let spriteCache = ImageCache()
+    // Caching the single legacy shield icon images.
+    let legacyCache = ImageCache()
+    // Caching the metadata info for Sprite.
     let infoCache =  SpriteInfoCache()
     var styleURI: StyleURI = .navigationDay
     var baseURL: URL = URL(string: "https://api.mapbox.com/styles/v1")!
@@ -15,41 +19,94 @@ class SpriteRepository {
             imageDownloader = ImageDownloader(sessionConfiguration: sessionConfiguration)
         }
     }
+    
+    func updateStyle(styleURI: StyleURI,
+                     representation: VisualInstruction.Component.ImageRepresentation? = nil,
+                     completion: @escaping CompletionHandler) {
+        guard styleURI != self.styleURI else {
+            completion()
+            return
+        }
+        
+        // Reset the default styleURI when style changes without valid ImageRepresentation.
+        guard let baseURL = representation?.shield?.baseURL else {
+            self.styleURI = styleURI
+            completion()
+            return
+        }
+        
+        // Reset Sprite cache when style changes with valid ImageRepresentation.
+        updateSprite(styleURI: styleURI, baseURL: baseURL, completion: completion)
+    }
 
-    func updateRepository(styleURI: StyleURI? = nil, representation: VisualInstruction.Component.ImageRepresentation? = nil, completion: @escaping CompletionHandler) {
+    func updateRepresentation(representation: VisualInstruction.Component.ImageRepresentation? = nil, completion: @escaping CompletionHandler) {
         let dispatchGroup = DispatchGroup()
 
-        // Reset cache and download the Sprite image and Sprite info only when the map style changes or the shield baseURL changes.
-        if (styleURI != self.styleURI) || (representation?.shield?.baseURL != self.baseURL) {
-            resetCache()
-            let styleURI = styleURI ?? self.styleURI
-            let baseURL = representation?.shield?.baseURL ?? self.baseURL
-            
-            if let styleID = styleURI.rawValue.components(separatedBy: "styles")[safe: 1],
-               let infoRequestURL = spriteURL(isImage: false, baseURL: baseURL, styleID: styleID),
-               let spriteRequestURL = spriteURL(isImage: true, baseURL: baseURL, styleID: styleID) {
-                
-                dispatchGroup.enter()
-                downloadInfo(infoRequestURL) { (_) in
-                    dispatchGroup.leave()
-                }
-                
-                dispatchGroup.enter()
-                downloadSprite(spriteRequestURL) { (_) in
-                    self.styleURI = styleURI
-                    self.baseURL = baseURL
-                    dispatchGroup.leave()
-                }
+        // Reset Sprite cache when the baseURL changes or no valid Sprite image cached.
+        let baseURL = representation?.shield?.baseURL ?? self.baseURL
+        if baseURL != self.baseURL || (getSpriteImage() == nil) {
+            dispatchGroup.enter()
+            updateSprite(styleURI: self.styleURI, baseURL: baseURL) {
+                dispatchGroup.leave()
             }
         }
         
         dispatchGroup.enter()
-        downloadLegacyShield(representation: representation) { (_) in
+        updateLegacy(representation: representation) {
             dispatchGroup.leave()
         }
         
         dispatchGroup.notify(queue: .main) {
             completion()
+        }
+    }
+    
+    func updateSprite(styleURI: StyleURI, baseURL: URL, completion: @escaping CompletionHandler) {
+        spriteCache.clearMemory()
+        infoCache.clearMemory()
+        
+        // Update the styleURI and baseURL just after the Sprite memory reset. When the connection is poor, the next round of style update
+        // or the representation update could use the correct ones.
+        self.styleURI = styleURI
+        self.baseURL = baseURL
+        
+        guard let styleID = styleURI.rawValue.components(separatedBy: "styles")[safe: 1],
+              let infoRequestURL = spriteURL(isImage: false, baseURL: baseURL, styleID: styleID),
+              let spriteRequestURL = spriteURL(isImage: true, baseURL: baseURL, styleID: styleID) else {
+                  completion()
+                  return
+              }
+        
+        let spriteKey = "\(styleID)-\(baseURL.absoluteString)"
+        let dispatchGroup = DispatchGroup()
+        
+        dispatchGroup.enter()
+        downloadInfo(infoRequestURL, spriteKey: spriteKey) { (_) in
+            dispatchGroup.leave()
+        }
+        
+        dispatchGroup.enter()
+        downloadSprite(spriteRequestURL, spriteKey: spriteKey) { (_) in
+            dispatchGroup.leave()
+        }
+        
+        dispatchGroup.notify(queue: .main) {
+            completion()
+        }
+    }
+    
+    func updateLegacy(representation: VisualInstruction.Component.ImageRepresentation? = nil, completion: @escaping CompletionHandler) {
+        guard let cacheKey = representation?.legacyCacheKey else {
+            completion()
+            return
+        }
+        
+        if let _ = legacyCache.image(forKey: cacheKey) {
+            completion()
+        } else {
+            downloadLegacyShield(representation: representation) { (_) in
+                completion()
+            }
         }
     }
     
@@ -64,14 +121,14 @@ class SpriteRepository {
         return urlComponent.url
     }
     
-    func downloadInfo(_ infoURL: URL, completion: @escaping (Data?) -> Void) {
+    func downloadInfo(_ infoURL: URL, spriteKey: String, completion: @escaping (Data?) -> Void) {
         let _ = imageDownloader.downloadImage(with: infoURL, completion: { [weak self] (_, data, error)  in
             guard let strongSelf = self, let data = data else {
                 completion(nil)
                 return
             }
 
-            guard strongSelf.infoCache.store(data) else {
+            guard strongSelf.infoCache.store(data, spriteKey: spriteKey) else {
                 completion(nil)
                 return
             }
@@ -80,24 +137,25 @@ class SpriteRepository {
         })
     }
     
-    func downloadSprite(_ spriteURL: URL, completion: @escaping (UIImage?) -> Void) {
+    func downloadSprite(_ spriteURL: URL, spriteKey: String, completion: @escaping (UIImage?) -> Void) {
         let _ = imageDownloader.downloadImage(with: spriteURL, completion: { [weak self] (image, data, error) in
             guard let strongSelf = self, let image = image else {
                 completion(nil)
                 return
             }
 
-            strongSelf.imageCache.store(image, forKey: "Sprite", toDisk: false, completion: {
+            strongSelf.spriteCache.store(image, forKey: spriteKey, toDisk: false, completion: {
                 completion(image)
             })
         })
     }
     
     func downloadLegacyShield(representation: VisualInstruction.Component.ImageRepresentation? = nil, completion: @escaping (UIImage?) -> Void) {
-        guard let legacyURL = representation?.imageURL(scale: VisualInstruction.Component.scale, format: .png) else {
-            completion(nil)
-            return
-        }
+        guard let legacyURL = representation?.imageURL(scale: VisualInstruction.Component.scale, format: .png),
+              let cacheKey = representation?.legacyCacheKey else {
+                  completion(nil)
+                  return
+              }
         
         let _ = imageDownloader.downloadImage(with: legacyURL, completion: { [weak self] (image, data, error) in
             guard let strongSelf = self, let image = image else {
@@ -105,30 +163,34 @@ class SpriteRepository {
                 return
             }
 
-            strongSelf.imageCache.store(image, forKey: "Legacy", toDisk: false, completion: {
+            strongSelf.legacyCache.store(image, forKey: cacheKey, toDisk: false, completion: {
                 completion(image)
             })
         })
     }
     
-    func getShield(displayRef: String, name: String) -> UIImage? {
-        let iconLeght = (displayRef.count < 2 ) ? 2 : displayRef.count
-        let infoName = name + "-\(iconLeght)"
+    func getShieldIcon(shield: VisualInstruction.Component.ShieldRepresentation) -> UIImage? {
+        // Use the styleURI and baseURL of current repository for spriteKey.
+        guard let styleID = styleURI.rawValue.components(separatedBy: "styles")[safe: 1] else { return nil }
+        let spriteKey = "\(styleID)-\(shield.baseURL.absoluteString)"
+
+        let iconLeght = max(shield.text.count, 2)
+        let shieldKey = shield.name + "-\(iconLeght)" + "-\(spriteKey)"
         
         // Retrieve the single shield icon if it has been cached.
-        if let shieldIcon = imageCache.image(forKey: infoName) {
+        if let shieldIcon = spriteCache.image(forKey: shieldKey) {
             return shieldIcon
         }
         
-        guard let spriteImage = imageCache.image(forKey: "Sprite"),
-              let spriteInfo = infoCache.spriteInfo(forKey: infoName) else { return nil }
+        guard let spriteImage = spriteCache.image(forKey: spriteKey),
+              let spriteInfo = infoCache.spriteInfo(forKey: shieldKey) else { return nil }
         
         let shieldRect = CGRect(x: spriteInfo.x, y: spriteInfo.y, width: spriteInfo.width, height: spriteInfo.height)
         if let croppedCGIImage = spriteImage.cgImage?.cropping(to: shieldRect) {
             
             // Cache the single shield icon if it hasn't been stored.
             let shieldIcon = UIImage(cgImage: croppedCGIImage)
-            imageCache.store(shieldIcon, forKey: infoName, toDisk: false, completion: nil)
+            spriteCache.store(shieldIcon, forKey: shieldKey, toDisk: false, completion: nil)
             
             return UIImage(cgImage: croppedCGIImage)
         }
@@ -136,13 +198,21 @@ class SpriteRepository {
         return nil
     }
     
-    func getLegacyShield() -> UIImage? {
-        return imageCache.image(forKey: "Legacy")
+    func getLegacyShield(with cacheKey: String?) -> UIImage? {
+        return legacyCache.image(forKey: cacheKey)
+    }
+    
+    func getSpriteImage() -> UIImage? {
+        // Use the styleURI and baseURL of current repository to retrieve Sprite image.
+        guard let styleID = styleURI.rawValue.components(separatedBy: "styles")[safe: 1] else { return nil }
+        let spriteKey = "\(styleID)-\(baseURL.absoluteString)"
+        return spriteCache.image(forKey: spriteKey)
     }
     
     func resetCache() {
-        imageCache.clearMemory()
+        spriteCache.clearMemory()
         infoCache.clearMemory()
+        legacyCache.clearMemory()
     }
 
 }